Latest CHE

+++ wsmaster/che-core-api-workspace/src/main/java/org/eclipse/che/api/workspace/server/TemporaryWorkspaceRemover.java
@@ -1,9 +1,10 @@
 /*
- * Copyright (c) 2012-2017 Red Hat, Inc.
- * All rights reserved. This program and the accompanying materials
- * are made available under the terms of the Eclipse Public License v1.0
- * which accompanies this distribution, and is available at
- * http://www.eclipse.org/legal/epl-v10.html
+ * Copyright (c) 2012-2018 Red Hat, Inc.
+ * This program and the accompanying materials are made
+ * available under the terms of the Eclipse Public License 2.0
+ * which is available at https://www.eclipse.org/legal/epl-2.0/
+ *
+ * SPDX-License-Identifier: EPL-2.0
  *
  * Contributors:
  *   Red Hat, Inc. - initial API and implementation
@@ -14,17 +15,17 @@ import static org.slf4j.LoggerFactory.getLogger;
 
 import com.google.common.annotations.VisibleForTesting;
 import javax.annotation.PostConstruct;
-import javax.annotation.PreDestroy;
 import javax.inject.Inject;
 import javax.inject.Singleton;
 import org.eclipse.che.api.core.Pages;
 import org.eclipse.che.api.core.ServerException;
+import org.eclipse.che.api.core.model.workspace.WorkspaceStatus;
 import org.eclipse.che.api.workspace.server.model.impl.WorkspaceImpl;
 import org.eclipse.che.api.workspace.server.spi.WorkspaceDao;
 import org.slf4j.Logger;
 
 /**
- * Removes temporary workspaces on server startup and shutdown.
+ * Removes stopped temporary workspaces on server startup and shutdown.
  *
  * @author Max Shaposhnik (mshaposhnik@codenvy.com)
  */
@@ -34,10 +35,12 @@ public class TemporaryWorkspaceRemover {
   private static final Logger LOG = getLogger(TemporaryWorkspaceRemover.class);
 
   private final WorkspaceDao workspaceDao;
+  private final WorkspaceRuntimes runtimes;
 
   @Inject
-  public TemporaryWorkspaceRemover(WorkspaceDao workspaceDao) {
+  public TemporaryWorkspaceRemover(WorkspaceDao workspaceDao, WorkspaceRuntimes runtimes) {
     this.workspaceDao = workspaceDao;
+    this.runtimes = runtimes;
   }
 
   @PostConstruct
@@ -45,16 +48,15 @@ public class TemporaryWorkspaceRemover {
     try {
       removeTemporaryWs();
     } catch (ServerException e) {
-      LOG.warn("Unable to cleanup temporary workspaces on startup: " + e.getLocalizedMessage(), e);
+      LOG.warn("Unable to cleanup temporary workspaces on startup: " + e.getMessage(), e);
     }
   }
 
-  @PreDestroy
   void shutdown() {
     try {
       removeTemporaryWs();
     } catch (ServerException e) {
-      LOG.warn("Unable to cleanup temporary workspaces on shutdown: " + e.getLocalizedMessage(), e);
+      LOG.warn("Unable to cleanup temporary workspaces on shutdown: " + e.getMessage(), e);
     }
   }
 
@@ -63,13 +65,16 @@ public class TemporaryWorkspaceRemover {
     for (WorkspaceImpl workspace :
         Pages.iterate(
             (maxItems, skipCount) -> workspaceDao.getWorkspaces(true, maxItems, skipCount))) {
-      try {
-        workspaceDao.remove(workspace.getId());
-      } catch (ServerException e) {
-        LOG.error(
-            "Unable to cleanup temporary workspace {}. Reason is {}",
-            workspace.getId(),
-            e.getLocalizedMessage());
+      WorkspaceStatus status = runtimes.getStatus(workspace.getId());
+      if (status == WorkspaceStatus.STOPPED) {
+        try {
+          workspaceDao.remove(workspace.getId());
+        } catch (ServerException e) {
+          LOG.error(
+              "Unable to cleanup temporary workspace {}. Reason is {}",
+              workspace.getId(),
+              e.getMessage());
+        }
       }
     }
   }
